Update README.md
Updated readme to show that this project is no longer active and that the codebase has been moved to gitlab.  

It might also be worth looking at https://docs.gitlab.com/ee/user/project/repository/repository_mirroring.html to just mirror the gitlab to here to avoid confusion.  with cloning this repo and having issues with the code not working.